secuirty fixes and layout fixes
git-svn-id: http://plugins.svn.wordpress.org/sliding-youtube-gallery/trunk@653956 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ engine/SygPlugin.php
@@ -382,23 +382,26 @@ class SygPlugin extends SanityPluginFramework {
 	 */
 	public static function uninstall() {
 		global $wpdb;
-
-		// get table name
-		$galleryTable = $wpdb->prefix . 'syg';
-		$styleTable = $wpdb->prefix . 'syg_OLD_V12X';
-		$backupTable = $wpdb->prefix . 'syg_styles';
-
-		// remove table
-		$dao = new SygDao();
-		$dao->removeTable($galleryTable);
-		$dao->removeTable($styleTable);
-		$dao->removeTable($backupTable);
-
-		// remove options
-		delete_option("syg_db_version");
-
-		// send stat
-		self::notify(SygConstant::BE_ACTION_UNINSTALL);
+		
+		// check if admin
+		if (is_admin()) {
+			// get table name
+			$galleryTable = $wpdb->prefix . 'syg';
+			$styleTable = $wpdb->prefix . 'syg_OLD_V12X';
+			$backupTable = $wpdb->prefix . 'syg_styles';
+	
+			// remove table
+			$dao = new SygDao();
+			$dao->removeTable($galleryTable);
+			$dao->removeTable($styleTable);
+			$dao->removeTable($backupTable);
+	
+			// remove options
+			delete_option("syg_db_version");
+	
+			// send stat
+			self::notify(SygConstant::BE_ACTION_UNINSTALL);
+		}
 	}
 
 	/**
@@ -407,8 +410,11 @@ class SygPlugin extends SanityPluginFramework {
 	 * @since 1.0.1
 	 */
 	public static function deactivation() {
-		// send stat
-		self::notify(SygConstant::BE_ACTION_DEACTIVATION);
+		// check if admin
+		if (is_admin()) {
+			// send stat
+			self::notify(SygConstant::BE_ACTION_DEACTIVATION);
+		}
 	}
 
 	/**
@@ -417,7 +423,10 @@ class SygPlugin extends SanityPluginFramework {
 	 * @since 1.0.1
 	 */
 	public static function activation() {
-		self::checkUpdateProcess();
+		// check if admin
+		if (is_admin()) {
+			self::checkUpdateProcess();
+		}
 	}
 
 	/************************************/
@@ -987,168 +996,170 @@ class SygPlugin extends SanityPluginFramework {
 	 * @return $output
 	 */
 	public function forwardToSettings($updated = false) {
-		// if we've updated a record set action to null
-		$action = ($updated == true) ? 'redirect' : $_GET['action'];
-		// determine wich action to call
-		switch ($action) {
-			case 'cache':
-				// update cached gallery
-				$galleries = $this->sygDao->getAllCachedGallery();
-				for ($i=0; $i<count($galleries); $i++) {
-					$gallery = $galleries[$i];
-					$gallery->cacheGallery();
-				}
-				break;
-			default:
-				if (isset($_POST['syg_submit_hidden'])
-						&& $_POST['syg_submit_hidden'] == 'Y') {
-					// database add/edit settings procedure
-					// get posted values
-					$data = serialize($_POST);
-					try {
-						// validate data
-						$valid = SygValidate::validateSettings($data);
-				
-						(get_option('syg_option_apikey') === false) ? add_option(
-								'syg_option_apikey',
-								$_POST['syg_option_apikey'])
-								: update_option('syg_option_apikey',
-										$_POST['syg_option_apikey']);
-				
-						(get_option('syg_option_which_thumb') === false) ? add_option(
-								'syg_option_which_thumb',
-								$_POST['syg_option_which_thumb'])
-								: update_option('syg_option_which_thumb',
-										$_POST['syg_option_which_thumb']);
-				
-						(get_option('syg_option_numrec') === false) ? add_option(
-								'syg_option_numrec',
-								$_POST['syg_option_numrec'])
-								: update_option('syg_option_numrec',
-										$_POST['syg_option_numrec']);
-				
-						(get_option('syg_option_pagenumrec') === false) ? add_option(
-								'syg_option_pagenumrec',
-								$_POST['syg_option_pagenumrec'])
-								: update_option('syg_option_pagenumrec',
-										$_POST['syg_option_pagenumrec']);
-						(get_option('syg_option_paginationarea') === false) ? add_option(
-								'syg_option_paginationarea',
-								$_POST['syg_option_paginationarea'])
-								: update_option('syg_option_paginationarea',
-										$_POST['syg_option_paginationarea']);
-				
-						(get_option('syg_option_paginator_borderradius') === false) ? add_option(
-								'syg_option_paginator_borderradius',
-								$_POST['syg_option_paginator_borderradius'])
-								: update_option('syg_option_paginator_borderradius',
-										$_POST['syg_option_paginator_borderradius']);
-						(get_option('syg_option_paginator_bordersize') === false) ? add_option(
-								'syg_option_paginator_bordersize',
-								$_POST['syg_option_paginator_bordersize'])
-								: update_option('syg_option_paginator_bordersize',
-										$_POST['syg_option_paginator_bordersize']);
-						(get_option('syg_option_paginator_bordercolor') === false) ? add_option(
-								'syg_option_paginator_bordercolor',
-								$_POST['syg_option_paginator_bordercolor'])
-								: update_option('syg_option_paginator_bordercolor',
-										$_POST['syg_option_paginator_bordercolor']);
-						(get_option('syg_option_paginator_bgcolor') === false) ? add_option(
-								'syg_option_paginator_bgcolor',
-								$_POST['syg_option_paginator_bgcolor'])
-								: update_option('syg_option_paginator_bgcolor',
-										$_POST['syg_option_paginator_bgcolor']);
-						(get_option('syg_option_paginator_shadowcolor') === false) ? add_option(
-								'syg_option_paginator_shadowcolor',
-								$_POST['syg_option_paginator_shadowcolor'])
-								: update_option('syg_option_paginator_shadowcolor',
-										$_POST['syg_option_paginator_shadowcolor']);
-						(get_option('syg_option_paginator_shadowsize') === false) ? add_option(
-								'syg_option_paginator_shadowsize',
-								$_POST['syg_option_paginator_shadowsize'])
-								: update_option('syg_option_paginator_shadowsize',
-										$_POST['syg_option_paginator_shadowsize']);
-						(get_option('syg_option_paginator_fontcolor') === false) ? add_option(
-								'syg_option_paginator_fontcolor',
-								$_POST['syg_option_paginator_fontcolor'])
-								: update_option('syg_option_paginator_fontcolor',
-										$_POST['syg_option_paginator_fontcolor']);
-						(get_option('syg_option_paginator_fontsize') === false) ? add_option(
-								'syg_option_paginator_fontsize',
-								$_POST['syg_option_paginator_fontsize'])
-								: update_option('syg_option_paginator_fontsize',
-										$_POST['syg_option_paginator_fontsize']);
-						// 3d carousel section
-						(get_option('syg_option_carousel_autorotate') === false) ? add_option(
-								'syg_option_carousel_autorotate',
-								$_POST['syg_option_carousel_autorotate'])
-								: update_option('syg_option_carousel_autorotate',
-										$_POST['syg_option_carousel_autorotate']);
-						(get_option('syg_option_carousel_delay') === false) ? add_option(
-								'syg_option_carousel_delay',
-								$_POST['syg_option_carousel_delay'])
-								: update_option('syg_option_carousel_delay',
-										$_POST['syg_option_carousel_delay']);
-						(get_option('syg_option_carousel_fps') === false) ? add_option(
-								'syg_option_carousel_fps',
-								$_POST['syg_option_carousel_fps'])
-								: update_option('syg_option_carousel_fps',
-										$_POST['syg_option_carousel_fps']);
-						(get_option('syg_option_carousel_speed') === false) ? add_option(
-								'syg_option_carousel_speed',
-								$_POST['syg_option_carousel_speed'])
-								: update_option('syg_option_carousel_speed',
-										$_POST['syg_option_carousel_speed']);
-						(get_option('syg_option_carousel_minscale') === false) ? add_option(
-								'syg_option_carousel_minscale',
-								$_POST['syg_option_carousel_minscale'])
-								: update_option('syg_option_carousel_minscale',
-										$_POST['syg_option_carousel_minscale']);
-						(get_option('syg_option_carousel_reflheight') === false) ? add_option(
-								'syg_option_carousel_reflheight',
-								$_POST['syg_option_carousel_reflheight'])
-								: update_option('syg_option_carousel_reflheight',
-										$_POST['syg_option_carousel_reflheight']);
-						(get_option('syg_option_carousel_reflgap') === false) ? add_option(
-								'syg_option_carousel_reflgap',
-								$_POST['syg_option_carousel_reflgap'])
-								: update_option('syg_option_carousel_reflgap',
-										$_POST['syg_option_carousel_reflgap']);
-						(get_option('syg_option_carousel_reflopacity') === false) ? add_option(
-								'syg_option_carousel_reflopacity',
-								$_POST['syg_option_carousel_reflopacity'])
-								: update_option('syg_option_carousel_reflopacity',
-										$_POST['syg_option_carousel_reflopacity']);
-						$this->data['redirect_url'] = '?page='
-						. SygConstant::BE_ACTION_MANAGE_SETTINGS . '&modified=true';
-						
-						// render adminStyles view
-						return $this->render('redirect');
-					} catch (SygValidateException $ex) {
-						// set the error
-						$this->data['exception'] = true;
-						$this->data['exception_message'] = $ex->getMessage();
-						$this->data['exception_detail'] = $ex->getProblems();
-					} catch (Exception $ex) {
-						// set the error
-						$this->data['exception'] = true;
-						$this->data['exception_message'] = $ex->getMessage();
+		if (is_admin()) {
+			// if we've updated a record set action to null
+			$action = ($updated == true) ? 'redirect' : $_GET['action'];
+			// determine wich action to call
+			switch ($action) {
+				case 'cache':
+					// update cached gallery
+					$galleries = $this->sygDao->getAllCachedGallery();
+					for ($i=0; $i<count($galleries); $i++) {
+						$gallery = $galleries[$i];
+						$gallery->cacheGallery();
 					}
-				}
-				break;	
+					break;
+				default:
+					if (isset($_POST['syg_submit_hidden'])
+							&& $_POST['syg_submit_hidden'] == 'Y') {
+						// database add/edit settings procedure
+						// get posted values
+						$data = serialize($_POST);
+						try {
+							// validate data
+							$valid = SygValidate::validateSettings($data);
+					
+							(get_option('syg_option_apikey') === false) ? add_option(
+									'syg_option_apikey',
+									$_POST['syg_option_apikey'])
+									: update_option('syg_option_apikey',
+											$_POST['syg_option_apikey']);
+					
+							(get_option('syg_option_which_thumb') === false) ? add_option(
+									'syg_option_which_thumb',
+									$_POST['syg_option_which_thumb'])
+									: update_option('syg_option_which_thumb',
+											$_POST['syg_option_which_thumb']);
+					
+							(get_option('syg_option_numrec') === false) ? add_option(
+									'syg_option_numrec',
+									$_POST['syg_option_numrec'])
+									: update_option('syg_option_numrec',
+											$_POST['syg_option_numrec']);
+					
+							(get_option('syg_option_pagenumrec') === false) ? add_option(
+									'syg_option_pagenumrec',
+									$_POST['syg_option_pagenumrec'])
+									: update_option('syg_option_pagenumrec',
+											$_POST['syg_option_pagenumrec']);
+							(get_option('syg_option_paginationarea') === false) ? add_option(
+									'syg_option_paginationarea',
+									$_POST['syg_option_paginationarea'])
+									: update_option('syg_option_paginationarea',
+											$_POST['syg_option_paginationarea']);
+					
+							(get_option('syg_option_paginator_borderradius') === false) ? add_option(
+									'syg_option_paginator_borderradius',
+									$_POST['syg_option_paginator_borderradius'])
+									: update_option('syg_option_paginator_borderradius',
+											$_POST['syg_option_paginator_borderradius']);
+							(get_option('syg_option_paginator_bordersize') === false) ? add_option(
+									'syg_option_paginator_bordersize',
+									$_POST['syg_option_paginator_bordersize'])
+									: update_option('syg_option_paginator_bordersize',
+											$_POST['syg_option_paginator_bordersize']);
+							(get_option('syg_option_paginator_bordercolor') === false) ? add_option(
+									'syg_option_paginator_bordercolor',
+									$_POST['syg_option_paginator_bordercolor'])
+									: update_option('syg_option_paginator_bordercolor',
+											$_POST['syg_option_paginator_bordercolor']);
+							(get_option('syg_option_paginator_bgcolor') === false) ? add_option(
+									'syg_option_paginator_bgcolor',
+									$_POST['syg_option_paginator_bgcolor'])
+									: update_option('syg_option_paginator_bgcolor',
+											$_POST['syg_option_paginator_bgcolor']);
+							(get_option('syg_option_paginator_shadowcolor') === false) ? add_option(
+									'syg_option_paginator_shadowcolor',
+									$_POST['syg_option_paginator_shadowcolor'])
+									: update_option('syg_option_paginator_shadowcolor',
+											$_POST['syg_option_paginator_shadowcolor']);
+							(get_option('syg_option_paginator_shadowsize') === false) ? add_option(
+									'syg_option_paginator_shadowsize',
+									$_POST['syg_option_paginator_shadowsize'])
+									: update_option('syg_option_paginator_shadowsize',
+											$_POST['syg_option_paginator_shadowsize']);
+							(get_option('syg_option_paginator_fontcolor') === false) ? add_option(
+									'syg_option_paginator_fontcolor',
+									$_POST['syg_option_paginator_fontcolor'])
+									: update_option('syg_option_paginator_fontcolor',
+											$_POST['syg_option_paginator_fontcolor']);
+							(get_option('syg_option_paginator_fontsize') === false) ? add_option(
+									'syg_option_paginator_fontsize',
+									$_POST['syg_option_paginator_fontsize'])
+									: update_option('syg_option_paginator_fontsize',
+											$_POST['syg_option_paginator_fontsize']);
+							// 3d carousel section
+							(get_option('syg_option_carousel_autorotate') === false) ? add_option(
+									'syg_option_carousel_autorotate',
+									$_POST['syg_option_carousel_autorotate'])
+									: update_option('syg_option_carousel_autorotate',
+											$_POST['syg_option_carousel_autorotate']);
+							(get_option('syg_option_carousel_delay') === false) ? add_option(
+									'syg_option_carousel_delay',
+									$_POST['syg_option_carousel_delay'])
+									: update_option('syg_option_carousel_delay',
+											$_POST['syg_option_carousel_delay']);
+							(get_option('syg_option_carousel_fps') === false) ? add_option(
+									'syg_option_carousel_fps',
+									$_POST['syg_option_carousel_fps'])
+									: update_option('syg_option_carousel_fps',
+											$_POST['syg_option_carousel_fps']);
+							(get_option('syg_option_carousel_speed') === false) ? add_option(
+									'syg_option_carousel_speed',
+									$_POST['syg_option_carousel_speed'])
+									: update_option('syg_option_carousel_speed',
+											$_POST['syg_option_carousel_speed']);
+							(get_option('syg_option_carousel_minscale') === false) ? add_option(
+									'syg_option_carousel_minscale',
+									$_POST['syg_option_carousel_minscale'])
+									: update_option('syg_option_carousel_minscale',
+											$_POST['syg_option_carousel_minscale']);
+							(get_option('syg_option_carousel_reflheight') === false) ? add_option(
+									'syg_option_carousel_reflheight',
+									$_POST['syg_option_carousel_reflheight'])
+									: update_option('syg_option_carousel_reflheight',
+											$_POST['syg_option_carousel_reflheight']);
+							(get_option('syg_option_carousel_reflgap') === false) ? add_option(
+									'syg_option_carousel_reflgap',
+									$_POST['syg_option_carousel_reflgap'])
+									: update_option('syg_option_carousel_reflgap',
+											$_POST['syg_option_carousel_reflgap']);
+							(get_option('syg_option_carousel_reflopacity') === false) ? add_option(
+									'syg_option_carousel_reflopacity',
+									$_POST['syg_option_carousel_reflopacity'])
+									: update_option('syg_option_carousel_reflopacity',
+											$_POST['syg_option_carousel_reflopacity']);
+							$this->data['redirect_url'] = '?page='
+							. SygConstant::BE_ACTION_MANAGE_SETTINGS . '&modified=true';
+							
+							// render adminStyles view
+							return $this->render('redirect');
+						} catch (SygValidateException $ex) {
+							// set the error
+							$this->data['exception'] = true;
+							$this->data['exception_message'] = $ex->getMessage();
+							$this->data['exception_detail'] = $ex->getProblems();
+						} catch (Exception $ex) {
+							// set the error
+							$this->data['exception'] = true;
+							$this->data['exception_message'] = $ex->getMessage();
+						}
+					}
+					break;	
+			}
+			
+			// prepare header
+			$this->prepareHeader($this->data, SygConstant::SYG_CTX_BE);
+	
+			// get settings
+			$options = $this->getOptions();
+	
+			// put settings in the view
+			$this->data['options'] = $options;
+	
+			// render adminSettings view
+			return $this->render('adminSettings');
 		}
-		
-		// prepare header
-		$this->prepareHeader($this->data, SygConstant::SYG_CTX_BE);
-
-		// get settings
-		$options = $this->getOptions();
-
-		// put settings in the view
-		$this->data['options'] = $options;
-
-		// render adminSettings view
-		return $this->render('adminSettings');
 	}
 
 	/**
@@ -1406,19 +1417,21 @@ class SygPlugin extends SanityPluginFramework {
 	 * @since 1.0.1
 	 */
 	public function forwardToDeleteGallery() {
-		// get the gallery id
-		$id = (int) $_GET['id'];
-		
-		// get the gallery
-		$gallery = $this->sygDao->getSygGalleryById($id);
-		
-		// remove file from cache
-		$gallery->removeFromCache();
-		
-		// delete gallery
-		$this->sygDao->deleteSygGallery($gallery);
-		
-		die();
+		if (current_user_can('delete_pages')) {
+			// get the gallery id
+			$id = (int) $_GET['id'];
+			
+			// get the gallery
+			$gallery = $this->sygDao->getSygGalleryById($id);
+			
+			// remove file from cache
+			$gallery->removeFromCache();
+			
+			// delete gallery
+			$this->sygDao->deleteSygGallery($gallery);
+			
+			die();
+		}
 	}
 	
 	/**
@@ -1427,16 +1440,18 @@ class SygPlugin extends SanityPluginFramework {
 	 * @since 1.4.0
 	 */
 	public function forwardToCacheGallery() {
-		// get the gallery id
-		$id = (int) $_GET['id'];
-		
-		// get the gallery
-		$gallery = $this->sygDao->getSygGalleryById($id);
-		
-		// cache gallery
-		$gallery->cacheGallery();
-		
-		die();
+		if (current_user_can('edit_posts')) {
+			// get the gallery id
+			$id = (int) $_GET['id'];
+			
+			// get the gallery
+			$gallery = $this->sygDao->getSygGalleryById($id);
+			
+			// cache gallery
+			$gallery->cacheGallery();
+			
+			die();
+		}
 	}
 
 	/**
@@ -1445,13 +1460,15 @@ class SygPlugin extends SanityPluginFramework {
 	 * @since 1.0.1
 	 */
 	public function forwardToDeleteStyle() {
-		// get the gallery id
-		$id = (int) $_GET['id'];
-
-		// delete gallery
-		$this->sygDao->deleteSygStyle($this->sygDao->getSygStyleById($id));
-
-		die();
+		if (current_user_can('delete_pages')) {
+			// get the gallery id
+			$id = (int) $_GET['id'];
+	
+			// delete gallery
+			$this->sygDao->deleteSygStyle($this->sygDao->getSygStyleById($id));
+	
+			die();
+		}
 	}
 
 	/**
@@ -1461,11 +1478,13 @@ class SygPlugin extends SanityPluginFramework {
 	 * @return $output
 	 */
 	public function forwardToSupport() {
-		// prepare header
-		$this->prepareHeader($this->data, SygConstant::SYG_CTX_BE);
-
-		// render contact view
-		return $this->render('adminSupport');
+		if (current_user_can('edit_posts')) {
+			// prepare header
+			$this->prepareHeader($this->data, SygConstant::SYG_CTX_BE);
+	
+			// render contact view
+			return $this->render('adminSupport');
+		}
 	}
 
 	/******************************/
